Inject dependencies directly into hook providers instead of using service locators.
Although the services aren't lazily instantiated, this simplifies the logic and
will make the hook providers easier to test.

The request handler still receives a service locator containing routes so those
can be instantiated lazily. It's usage is isolated to a single method that can
be stubbed out for tests.