Add a has_package() method to the package manager.
This fixes a bug causing all pending updates to be cached whether they're
whitelisted or not.

PackageManager::get_package() is called in at least one place to determine
whether a package is whitelisted, but it doesn't check the whitelist.

PackageManager::get_packages() is called in a couple of places, but the calling
code needs to check for the existence of a key.