Refactor packages (#52)
* Refactor package type creation.

This differentiates between basic package descriptions and installed packages
using the base Package interface and an InstalledPackage interface. Local
filesystem state has been moved out of the Plugin and Theme classes and is
instead injected using the builder pattern.

The package builders use reflection to set protected properties on package
instances to ensure packages are immutable once they're built.

Benefits of this approach are:

The package classes are no longer tied to the environment (WP constants, files,
API, etc). The builders should contain the only references to the environment
using from_source() convenience methods, which should demonstrate how to use the
builder interface.

Package instances should be immutable.

In the future, virtual packages could be created using the same builder.

* Introduce package repositories.

Repositories serve as an interface for working with packages as collections and
replace the package manager to further separate concerns.

This introduces a few repositories:

- InstalledPlugins/InstalledThemes - Wrap WordPress APIs to retrieve plugins and
  themes as packages.
- MultiRepository - Accepts an array of repository instances to work with them
  as if they were one.
- Whitelist - Implements the user-defined whitelist rules.
- Composer - Consumes another repository and transforms package data to the
  format expected in packages.json.
- CachedRepository - Wraps another repository to create an in-memory cache for
  the lifetime of the request after the repository has been initialized for the
  first time.

* Implement the new package and repository APIs.