Handle manual offsets when creating DateTimeZone()
`get_option( 'timezone_string' )` is only populated if the Settings -> General timezone field is set to a location. The field could also be set to a manual offset, like `UTC+2`, which leaves the `get_option( 'timezone_string' )` and the `get_option( 'gmt_offset' )` populated.

This addition allows the offset to be used when creating a `DateTimeZone` object for the API Key creation process.

Fixes #60.