Attempt to archive any missing release in the download route.
The download route previously only attempted to archive the currently installed
version, however the release manager only archives the installed version and
releases with a source URL (pending updates), both of which should already be
available by the time the request passes all the previous checks.

The release manager also checks to see if the release artifact already exists,
so there's no need for the extra validation in the download route.