Remove the handle_error() method from authentication servers.
This creates a new AuthenticationException exception to transport authentication
errors, including headers and status code, from the authenticate() method
instead of using a handle_error() method.

If the current request is a REST request, the Authentication provider will
convert the exception to an error instead of allowing the exception to bubble
up.