Cache the installed version of a package if it doesn't exist when viewing packages.json. See #131
It's possible to upgrade a whitelisted package without having an artifact
generated (through the filesystem, etc). The download route attempts to archive
packages from source on demand if they don't exist. This updates the Composer
repository transformer to try to cache the installed version from source if the
artifact doesn't exist so it can be included in the packages.json endpoint.