Rework when cached releases are added to a package.
The cache was being searched for all plugins and themes instead of just
the plugins and themes that were being managed by SatisPress. It's not
noticeable or much of an issue when using local storage, but it starts
to add up quick when trying to implement a remote storage provider.

This only searches for cached releases when a plugin or theme is being
managed by SatisPress.